+++ history.md
@@ -4,6 +4,7 @@
 
 - [x] Support pytorch dist FasterCocoEvaluator
 - [x] Support torchvision CocoDetection Dataset as FasterCocoDetection
+- [x] Return COCO.showAnns https://github.com/MiXaiLL76/faster_coco_eval/issues/52
 
 ## v1.6.5